backend/requirements.txt
Django
psycopg
psycopg2
GitPython
